Bugfix: try-use-prebuilt-node-rdkafka.sh: fix npm-semver conflict
When there wasn't a `semver` binary in ${PATH}, `npx semver` would
download the npm semver package and run it. Now, with `semver-tool`
installed at `/usr/local/semver`, `npx semver` (very helpfully) decides
that you meant to use the one in your ${PATH}.

Great. Very expected behavior, as usual, npm.

So, we need to explicitly install the npm semver package and invoke it
explicitly, so as to disambiguate it from semver-tool.