Pin cached node-rdkafka version in build-node image
PLAT-1327

`node-rdkafka@2.1.0` upgrades to `librdkafka@1.5.2` which since `1.5.0`
has deprecated the behavior where a Kafka consumer auto-creates a topic
if it doesn't exist when the consumer connects to Kafka. The integration
tests in `@better/events` and `@better/event-definitions` depend on this
behavior the way the tests are set up. The desired behavior is available
in `node-rdkafka@2.9.1` so pinning the version of the cached library to
`2.9.1`