Repair an error with name parsing in the build system
In this commit, I repair an error with the name parser for image names. The
current system simply splits the input on dashes, then takes certain elements.
This is a problem, because some images use dashes in their name, and this breaks
that.

In this commit, the new code now only splits on the *first* dash, so the rest of
the input becomes the image name. As a result, `make
build-build-python-postgres` now works.